+++ single.php
@@ -6,24 +6,29 @@
  *
  * @package     WordPress
  * @subpackage  MGPress
- * @version     1.0.1
+ * @version     1.0.2
  * @since       MGPress 1.0
  * @author      kchevalier@mindgruve.com
  */
 
 // set context
-$context         = Timber::get_context();
-$post            = Timber::query_post();
-$context['post'] = $post;
-
-// custom template
-$templateName = $post->meta('_wp_page_template') ? $post->meta('_wp_page_template') : null;
+$context = Timber::get_context();
 
 // controller (optional)
 $controller = dirname(__FILE__).'/controllers/detail-'.$post->post_type.'.php';
 if (file_exists($controller)) {
     include_once($controller);
 }
+
+// set post
+if (!isset($context['post'])) {
+    $post = Timber::query_post();
+}
+$context['post'] = $post;
+
+// custom template
+$templateName = $post->meta('_wp_page_template') ? $post->meta('_wp_page_template') : null;
+
 
 // determine templates
 if (post_password_required($post->ID)) {